Feat/course mangment (#10)
* feat : completed the required enums required for database

* feat : implemented the required entites requeird for the realted to course

attachemtns , courses , module , quizz

* feat: created the course service

essiental crud operation with basic filtering

* feat : implementd the crud operation endpoints for crud operations

* complete course crud operation

* fix : remove logs files

* feat/attachment: add file storage entity with user relations and metadata support

* feat/course: add course entity with enums, pricing, and module relationships

* feat/module: add module entity with video relationships and ordering support

* feat/video: add add/remove attachemnt for video

* feat/mapper: add attachment mapper with entity-DTO conversion and metadata handling

* feat/mapper: add course mapper with comprehensive entity-DTO conversion and update operations

* feat/mapper: add module mapper with entity-DTO conversion and video mapping support

* feat:mapper : add VideoMapper for converting between VideoEntity, VideoDto, and CreatVideoDto

* feat:dto : add AttachmentDto, AttachmentSummaryDto, CreateAttachmentDto, and UpdateAttachmentDto for attachment handling

* feat:dto : add CourseDetailsDto, CourseFilterDto, CourseModuleDto, CourseSummaryDto, CreateCourseDto, and UpdateCourseDto for course management

* feat:dto : add CreateModuleDto, DetailedModuleDto, ModuleSummaryDto, and UpdateModuleDto for module management

* feat:dto : add AddAttachmentToVideoDto, CreateAndAddAttachemntToVideoDto, CreatVideoDto, and VideoDto for video management

* feat:interface : add AttachmentService interface for managing attachment operations

* feat:interface : add CourseService interface for managing course operations

* feat:interface : add ModuleService interface for managing module operations

* feat:service : implement AttachmentServiceImpl for managing attachment CRUD and file handling

* feat:service : implement CourseServiceImpl for managing course CRUD, module association, filtering, and business validations

* feat:service : implement ModuleServiceImpl for managing module CRUD and video association

* feat/video:Add methods to attach and detach attachments from videos

* feat/attachment: add endpoints to create, update, delete, retrieve, and download attachments

* feat/course: add endpoints to manage courses and attach/remove modules

* feat/module-controller: add CRUD endpoints for modules and manage module videos

* completing course managment

* refactor(access-level): remove AccessLevel enum as it's no longer needed

* refactor/admin-controller: remove unused import for RolesName

* refactor(AuthController): remove unused import for PreAuthorize

* fix(CourseController): correct string formatting in success messages for module operations
fix(VideoController): optimize tag cleaning by using Stream API

* test(AttachmentServiceImpl): add unit tests for attachment service methods

* test(CourseServiceImpl): add unit tests for course service methods

* test(ModuleServiceImpl): add unit tests for module service methods

* test(S3Service): add unit tests for S3 service methods

* feat/Currency: introduce Currency enum and update related DTOs and entities

* fix/CreateCourseRequest: ensure currency field is required

* refactor/CourseServiceImpl: replace builder pattern with setter methods for course creation
fix/course_is_free: update the enitity,dot to only view it .

* fix/course_dtos :anotation validation fix

+++ src/main/java/com/example/e_learning_system/Repository/UserRepository.java
@@ -18,7 +18,7 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 
     public Optional<UserEntity> findByEmail(String email);
-    public Optional<UserEntity> findById(long id);
+    public Optional<UserEntity> findById(Integer id);
 
 
     public Page<UserEntity> findAll(Pageable pageable);
